html-proofer: Only report HTTP 4xx failures
This should make the tests way more stable but unfortunately also less
useful...

The current problem is that we're getting a lot of the following libcurl
error codes [0]:
- Couldn't resolve host name
- Couldn't connect to server
- SSL connect error (less common)

Some of these errors are valid but unfortunately most of them seem to be
related to the Travis CI infrastructure and cannot be reproduced
locally. They can e.g. be caused by problems with the local DNS resolver
or missing routes.

[0]: https://curl.haxx.se/libcurl/c/libcurl-errors.html