json2giggity.py: Keep the explicit line feeds
The behaviour should now be consistent with the Markdown files for the
website. However, I'm not sure if this really improves the situation.

In the landscape mode this should be fine, but for the portrait mode
this can cause issues, e.g. if the content is wrapped to ~70 characters
but only <70 characters fit in one line. Then the line will be wrapped
before the explicit line feed which will follow after a few additional
characters, resulting in two line feeds instead of one.

This is e.g. a problem for (not the fault of the speakers):
- klaus-knopper-cloudbusters-wer-kontrolliert-dein-smart-home.md
- michael-stroeder-openssh-certificates.md

But at the same time this is an improvement for:
- daniel-dreher-10-jahre-game-server-mit-linux.md
- uli-kleemann-buddel-die-eins-ueber-vpns-und-warum-man-das-selber-machen-sollte.md

The best solution would probably be to use lists instead of \n for
enumerations and only allow \n\n for new paragraphs.
TODO: We should add a note on cfp.tuebix.org that we use Markdown so
that the speakers know how the content will be formatted later (and
maybe generate a preview). Giggity seems to use HTML but with support
for the most common Markdown constructs (lists, headings, links, etc.).