bin: Cleanup the Giggity scripts
We used the XML export from Pretalx for 2023. These scripts threw
warning/error messages and I never got around to look into it. Using a
Pretalx link for Giggity should be better anyway as it updates
automatically and it seems to have worked fine for 2023. We might want
to alter the XML file a bit in the future but then it's probably best to
rewrite json2giggity.py from scratch (only using this version as a
reference for how the result should look like).