2023: Fix the URL IDs
Python's lstrip() method takes a **set** of characters to be removed
from the left side of the sting and therefore unsuitable for removing
prefixes. The documentation even explicitly mentions this [0]:
> The chars argument is not a prefix; rather, all combinations of its
> values are stripped
> See str.removeprefix() for a method that will remove a single prefix
> string rather than all of a set of characters.

Oh well...

Unfortunately, we repaired this way too late so we should definitely not
break the old links. Therefore, I came up with this quick and dirty
approach... (a better solution would be to generate redirects but that
can also be done later).

We also need to generate better URL IDs in the future as the new ones
are still problematic (especially due to the missing umlauts). I added a
note to convert_from_pretalx.py for a potential solution that gives us
the most flexibility (since the talk IDs from Pretalx are definitely
stable).
Anyway, let's use this quick fix for now.

[0]: https://docs.python.org/3.11/library/stdtypes.html#str.lstrip