schedule: Move the iOS app hint below the Android app hint
The order here did bug me.

It's nice that there is an Apple app but we should indicate that it's
unofficial.
Also: The source-code on GitHub [0] wasn't updated since version 1.6.1
in 2020 while version 2.0.7 is the current one (so there doesn't seem to
be any source-code available for the 8 most recent versions).

[0]: https://github.com/zenitaeglos/TuebixApp