bin/convert_from_pretalx: Abort with an error if /answers are paginated
This is the issue that I missed in fd1cac2.
The `limit` parameter doesn't exist anymore. There's a `page_size`
parameter now but it already defaults to the maximum of 50 [0]:
> By default, the page size (that is, the length of the results field)
> is 50. You can decrease the page size with the page_size parameter,
> but for performance reasons, you cannot increase it past 50.

Let's turn this into an error for now so that we don't miss it in the
future (we currently have 98 answers in total and 32 answers for
question 5 ("Vorwissen")).

[0]: https://docs.pretalx.org/api/fundamentals/#pagination