Fix the live reloading with a local `hugo serve`
Chromium reports the following CSP violation:
```
Refused to connect to 'ws://localhost:1313/livereload' because it
violates the following Content Security Policy directive: "default-src
'none'". Note that 'connect-src' was not explicitly set, so
'default-src' is used as a fallback.
```

This fixes the optional LiveReload functionality (at the cost of making
the CSP a bit less strict - we could alternatively only allow this for a
local `hugo serve` / dev setup).